Add response structures
- Added hit and source structures
- Renamed emailHandler to email_handler
- Changed module name
- Change handler name GetEmail to GetEmails
