+++ internal/models/email.go
@@ -1,0 +1,22 @@
+package models
+
+type Email struct {
+	MessageID               string   `json:"messageId"`
+	Date                    string   `json:"date"`
+	From                    string   `json:"from"`
+	To                      []string `json:"to"`
+	CC                      []string `json:"cc"`
+	BCC                     []string `json:"bcc"`
+	Subject                 string   `json:"subject"`
+	MimeVersion             string   `json:"mimeVersion"`
+	ContentType             string   `json:"contentType"`
+	ContentTransferEncoding string   `json:"contentTransferEncoding"`
+	XFrom                   string   `json:"xFrom"`
+	XTo                     []string `json:"xTo"`
+	Xcc                     []string `json:"xcc"`
+	Xbcc                    []string `json:"xbcc"`
+	XFolder                 string   `json:"xFolder"`
+	XOrigin                 string   `json:"xOrigin"`
+	XFileName               string   `json:"xFileName"`
+	Body                    string   `json:"body"`
+}
