+++ internal/models/hit.go
@@ -1,0 +1,10 @@
+package models
+
+type Hit struct {
+	Index     string  `json:"_index"`
+	Type      string  `json:"_type"`
+	Id        string  `json:"_id"`
+	Score     float64 `json:"_score"`
+	Timestamp string  `json:"@timestamp"`
+	Source    Source  `json:"_source"`
+}
