+++ internal/models/source.go
@@ -1,0 +1,7 @@
+package models
+
+type Source struct {
+	Email     Email  `json:"email"`
+	Timestamp string `json:"@timestamp"`
+	Path      string `json:"path"`
+}
